Safari-609.1.11.1/LayoutTests/http/tests/security/xss-DENIED-non-shadowable-propterty-with-iframe-proto.html

<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe src="http://localhost:8000/" style=""></iframe>
<pre id="console"></pre>
<script>
description("Tests that making other frame window a prototype doesn't expose that window properties");
jsTestIsAsync = true;

targetWindow = frames[0];

window.onload = function() {
  originalLocation = location;

  shouldThrowErrorName("__proto__ = targetWindow", "TypeError");

  shouldBeTrue('location === originalLocation');
  shouldBeTrue('this.location === originalLocation');
  finishJSTest();
}
</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
